requirements.txt
asgiref==3.5.2
attrs==21.4.0
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.5.1
PyJWT==2.4.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
whitenoise==6.1.0
